# Range formatting performance