--- a/lib/test/clojure_lsp/features/format_test.clj
+++ b/lib/test/clojure_lsp/features/format_test.clj
@@ -29,19 +29,41 @@
       (is (empty? r))
       (is (vector? r)))))
 
+(defn range-formatting [code]
+  (let [[[row col] [end-row end-col] :as positions] (h/load-code-and-locs code)]
+    (let [position-count (count positions)]
+       (assert (= 2 position-count) (format "Expected two cursors, got %s" position-count)))
+    (f.format/range-formatting (h/file-uri "file:///a.clj") {:row row :col col :end-row end-row :end-col end-col} db/db)))
+
 (deftest test-range-formatting
   (swap! db/db shared/deep-merge {:project-root-uri (h/file-uri "file:///project")})
-  (h/load-code-and-locs "(a  )\n(b c d)")
   (testing "when custom config file doesn't exists"
     (with-redefs [shared/file-exists? (constantly false)]
       (is (= [{:range {:start {:line 0 :character 0}
                        :end {:line 0 :character 5}}
                :new-text "(a)"}]
-             (f.format/range-formatting (h/file-uri "file:///a.clj") {:row 1 :col 1 :end-row 1 :end-col 4} db/db)))))
+             (range-formatting "|(a | )\n(b c d)")))))
   (testing "when custom config file exists"
     (with-redefs [shared/file-exists? (constantly true)
                   slurp (constantly "{}")]
       (is (= [{:range {:start {:line 0 :character 0}
                        :end {:line 0 :character 5}}
                :new-text "(a)"}]
-             (f.format/range-formatting (h/file-uri "file:///a.clj") {:row 1 :col 1 :end-row 1 :end-col 4} db/db))))))
+             (range-formatting "|(a | )\n(b c d)")))))
+  (testing "of several forms"
+    (is (= [{:range {:start {:line 0 :character 0}
+                     :end {:line 1 :character 5}}
+             :new-text "(a)\n(b)"}]
+           (range-formatting "(a | )\n|(b  )\n(c d)")))
+    (is (= [{:range {:start {:line 0 :character 0}
+                     :end {:line 1 :character 5}}
+             :new-text "(a)\n(b)"}]
+           (range-formatting "(a | )\n(|b  )\n(c d)")))
+    (is (= [{:range {:start {:line 0 :character 0}
+                     :end {:line 1 :character 0}}
+             :new-text "(a)\n"}]
+           (range-formatting "(a | )|\n(b  )\n(c d)")))
+    (is (= [{:range {:start {:line 0 :character 0}
+                     :end {:line 0 :character 5}}
+             :new-text "(a)"}]
+           (range-formatting "(a | |)\n(b  )\n(c d)")))))