This code investigates performance in range-formatting. Historically there have been reports (#266) of poor range formatting performance, particularly in large files for users who have auto-formatting turned on and who use paredit.

I haven't been able to reproduce really terrible performance. The worst I can do is a little over half a second to re-format large chunks of long files. For example when re-formatting the first half of a 2000 line file, which is mostly repeats of the line `(= (+ 1 2) 3)`, I get the following performance data:

```
original algorithm
614ms, std-dev 14ms
7% parsing
52% finding top forms
21% formatting
```

So that's odd... Why does it spend so much time finding the top-level forms? Let's look at `parser/find-top-forms-in-range`. First it finds every node in the range—every token, every sexp, everything. Then for _each one_ it searches upwards, to find a top-level node. Then it distincts those nodes. This is ... inefficient.

It would work equally well to search upward from the first node in the range, and the last node, and then take all the intervening top-level nodes. Let's try that.

```
first-and-last top-level
358ms, std-dev 5ms
10% parsing
9% finding top forms
43% formattting
```

This is better. (I think it also fixes some off-by-one bugs where the full range isn't formatted, but that's tangential.) Can we improve formatting time too? The original algorithm loops over the top level forms that cover the range, calling `cljfmt` for each one. But, perhaps it'd be faster to pass a single node that contains the top-level forms to `cljfmt`, so we only have to call it once.

```
first-and-last top-level, with merged formatting
301ms, std-dev 9ms
14% parsing
10% finding top forms
42% formatting
```

Not a huge improvement, but still better. And one last attempt... What if we use the more efficient cursor navigation from the `economy-of-motion` branch, which this branch is built on, to find the first and last top-level forms?

```
efficient first-and-last top-level, with merged formatting
271ms, std-dev 7ms
15% parsing
0.6% finding top forms
48% formatting
```

All together, that's about 50% faster, just by improving the `clojure-lsp` side. I still wouldn't call 271ms _fast_. I imagine it would be too slow for paredit users. But at least now most of the time is spent outside of `clojure-lsp`—in `rewrite-clj` or `cljfmt`. I think that's a win.

I'll clean this up and turn it into a real PR.