@mainej 2022-03-04
Github "helpfully" converts these interactive SVGs into static SVGs. I can provide the original interactive files if requested.

![01-range-formatting-orig](https://user-images.githubusercontent.com/6774/156689121-1f50d7fc-cc8f-414d-9434-5f88a24d6c38.svg)
![02-range-formatting-first-last](https://user-images.githubusercontent.com/6774/156689128-61b65a82-58ec-4ceb-90bc-26b90dc50715.svg)
![03-range-formatting-first-last-merge-fmt](https://user-images.githubusercontent.com/6774/156689131-5102121c-26dc-4317-87bb-9b84316d23ac.svg)
![04-range-formatting-to-loc-merge-fmt](https://user-images.githubusercontent.com/6774/156689133-114c01d5-d398-4186-b5ed-d80914877a36.svg)