@ericdallo 2022-03-04
Nice! the flamegraphs seems to help a lot on those performance overviews, I didn't think we could improve that much on clojure-lsp side, thank you!